backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
openpyxl
python-multipart
# python-dotenv will be needed for env vars
python-dotenv
google-generativeai>=0.5.0
openai
psycopg2-binary
